Quiet some warnings with OSX gcc
git-svn-id: http://svn.mythtv.org/svn/trunk@12417 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/frequencytables.cpp
@@ -7,7 +7,7 @@ freq_table_map_t frequencies;
 static void init_freq_tables(freq_table_map_t&);
 
 TransportScanItem::TransportScanItem()
-    : mplexid(-1),        FriendlyName(""),
+    : mplexid((uint)-1),  FriendlyName(""),
       friendlyNum(0),     SourceID(0),          UseTimer(false),
       scanning(false),    timeoutTune(1000)
 {
